refactor: Split recognition models into separate files
